pyp: simplify config parsing
Now that find_names implements scopes, we can basically just delegate to
it. This probably fixes a number of edge cases. The most salient
difference is probably that you can have multiple top-level lambda
assignments that share parameter names.

Previously, the conditional definition support we had was really basic,
but also pretty broken. E.g, we didn't visit the test for if statements
or the handlers for try statements. This fixes that, but there are still
issues with conditional definitions. Some of this falls down to
find_names, and some of this is that we usually want to err on the side
of false positives for defined names, but for top-level-ish conditional
definitions in config, we want to err on the side of false negatives.